% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frobenius.R
\name{dist_frobenius}
\alias{dist_frobenius}
\title{Frobenius Distance}
\usage{
dist_frobenius(graph_1, graph_2)
}
\arguments{
\item{graph_1}{igraph or matrix object.}

\item{graph_2}{igraph or matrix object.}
}
\value{
A numeric distance metric.
}
\description{
Frobenius norm between two adjacency matrices.
}
