% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsen-mikhailov.R
\name{dist_ipsen_mikhailov}
\alias{dist_ipsen_mikhailov}
\title{Ipsen Mikhailov}
\usage{
dist_ipsen_mikhailov(graph_1, graph_2, hwhm = 0.08, results_list = FALSE)
}
\arguments{
\item{graph_1}{igraph or matrix object.}

\item{graph_2}{igraph or matrix object.}

\item{hwhm}{Numeric parameter for the lorentzian kernel.}

\item{results_list}{Logical indicating whether or not to return results list.}
}
\description{
Ipsen Mikhailov
}
