% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{dist_jaccard}
\alias{dist_jaccard}
\title{Jaccard Distance}
\usage{
dist_jaccard(graph_1, graph_2)
}
\arguments{
\item{graph_1}{igraph or matrix object.}

\item{graph_2}{igraph or matrix object.}
}
\description{
Graph distance based on the Jaccard index between edge sets.
}
