% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian-spectral.R
\name{dist_laplacian_spectral}
\alias{dist_laplacian_spectral}
\title{Laplacian Spectral Distance}
\usage{
dist_laplacian_spectral(
  graph_1,
  graph_2,
  normed = TRUE,
  kernel = "normal",
  hwhm = 0.011775,
  measure = "jensen-shannon",
  k = NULL,
  which = "LM",
  results_list = FALSE
)
}
\arguments{
\item{graph_1}{igraph or matrix object.}

\item{graph_2}{igraph or matrix object.}

\item{normed}{Logical parameter for the lorentzian kernel.}

\item{kernel}{Character indicating kernel type. Can be "normal", "lorentzian", or NULL.}

\item{hwhm}{Numeric indicating half width at half maximum.}

\item{measure}{Character indicating measure type. Can be "jensen-shannon" or "euclidean".}

\item{k}{Numeric indicating number of eigenvalues kept, NULL means all.}

\item{which}{Character prioritizing which eigenvalues are kept, see get_eigvs below for more.}

\item{results_list}{Boolean indicating whether to returns
just distance if FALSE, list of additional info if TRUE}
}
\description{
Laplacian Spectral Distance
}
