% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial-dissimilarity.R
\name{dist_polynomial_dissimilarity}
\alias{dist_polynomial_dissimilarity}
\title{Polynomial Dissimilarity Distance}
\usage{
dist_polynomial_dissimilarity(graph_1, graph_2, k = 5, alpha = 1)
}
\arguments{
\item{graph_1}{igraph or matrix object.}

\item{graph_2}{igraph or matrix object.}

\item{k}{Numeric maximum degree of the polynomial.}

\item{alpha}{Numeric weighting factor.}
}
\value{
A numeric polynomial dissimilarity between graph_1, graph_2
  in a structure dist
}
\description{
Compares polynomials relating to the eigenvalues of the adjacency matrices
}
